opengl optimizations

+++ ps3emu/rsx/GLVARs.h
@@ -14,7 +14,7 @@ class GLVARs {
 
 public:
     inline void enable(int i) {
-        assert(i < _arr.size());
+        assert(i < (int)_arr.size());
         if (!_arr[i].enabled) {
             glEnableVertexAttribArray(i);
             _arr[i].enabled = true;
@@ -22,7 +22,7 @@ public:
     }
 
     inline void disable(int i) {
-        assert(i < _arr.size());
+        assert(i < (int)_arr.size());
         if (_arr[i].enabled) {
             glDisableVertexAttribArray(i);
             _arr[i].enabled = false;
@@ -30,7 +30,7 @@ public:
     }
 
     inline void setValue(int i, glm::vec4 const& value) {
-        assert(i < _arr.size());
+        assert(i < (int)_arr.size());
         if (_arr[i].value != value) {
             glVertexAttrib4f(i, value.r, value.g, value.b, value.a);
             _arr[i].value = value;
